Python-to-JS (P6)
Fix bgwriter_flush_after incorrect tuning on non-Windows platform
Merge every Python code into a single tuner.js
